Applying Jon's patch
